refactor(settings): split ProvidersTab into per-mode panels
Pure structural extraction — zero behavior change. The 1021-line
providers-tab.tsx collapses to 257 lines of layout + mode-card grid +
panel delegation; the three remaining mode panels (BYOK, ChatGPT
subscription, Custom endpoint) each move to their own file. Ollama and
Rapid-MLX panels were already extracted.

New files under frontend/src/components/settings/providers/:
- byok-panel.tsx              (215 lines) — BYOK provider list +
                                            updateProviderKey /
                                            deleteProviderKey /
                                            toggleProvider
- chatgpt-sub-panel.tsx       (246 lines) — OAuth login + polling +
                                            manual callback + disconnect
- custom-endpoint-panel.tsx   (228 lines) — Saved-endpoint list +
                                            delete/toggle + delete legacy
                                            local provider
- custom-endpoint-form.tsx    (120 lines) — Add-new form (own state)

State migration: each panel owns the input state and mutations scoped to
its concern. The parent passes only the data needed for panel rendering
(providers, localStatus, keyStatus). React Query cache deduplicates the
shared queries between parent (used for card-grid "connected" badges)
and child panels.

Frontend typecheck + ESLint clean. No translation key changes, no DOM
structure changes, no API contract changes.

Closes #102.
